Fixed button text and state logic
By convention the start button will now turn into a pause button when
you start. When you pause, the start button will be the resume button.
The stop button will always be a stop button. For your convenience the
stop button is now deactivated when the bot is stopped, and the settings
button is deactivated while the bot is running. Pausing or stopping the
bot will enable the settings button.