UI Update Race Fix
Even if the UI update task is canceled, there's a race where the UI
shows "Waiting for Fish (i/Ns)" instead of "Paused" or "Stopped".
Similarly, when the bot hears the fish, instead of "I hear a Fish!" the
number of seconds is displayed.

Here's the race condition for the cancellation scenario:

1/ UI update task schedules progress update on showing # of waiting
seconds
2/ cancel or pause button is hit
3/ UI update task is canceled*
4/ button triggers UI update to say 'canceled' or 'stopped'
5/ progress update triggers UI change to show "i/N seconds"

* technically this is where the cancellation is **requested**. there's
no guarantee that the task will actually get canceled here. however,
since the progress callback is on the same synchronization context as
the cancellation, this doesn't matter.

Thee race condition for the "I Hear a Fish!" scenario is basically the
same thing. Only difference is that instead of a UI button, the fish
sound will trigger cancellation, and instead of the cancellationToken,
the uiUpdateCancellationToken gets its IsCancellationRequested set to
true.

The progress callback happens on the same synchronization context that
the cancellation requests happen in, and those cancellation requests
precede their respective UI updates (for button-triggered or
fish-sound-triggered UI updates). Thus, when the progress callback
happens, there's no race on whether a cancellation was requested.

I could try to make cancellation wait for the UI task to finish before
updating the UI, but it's much easier instead to gate the progress
update callbacks this way.